BUG/MEDIUM: queue: prevent a backup server from draining the proxy's connections
When switching back from a backup to an active server, the backup server
currently continues to drain the proxy's connections, which is a problem
because it's not expected to be able to pick them.

This patch ensures that a backup server will only pick backend connections
if there is no active server and it is the selected backup server or all
backup servers are supposed to be used.

This issue seems to have existed forever, so this fix should be backported
to all stable versions.

(cherry picked from commit a8694654ba021bf1e0e560a98ab5e70dc44d212e)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 5550143cd6de58c6e733e389c6946e3dd26e89c0)
[wt: the logic was inverted between 1.7 and 1.8, in 1.7 we need to
 reverse all the condition as it consists in forcing to use the server
 connection instead of checking if we can pick a backend connection]
Signed-off-by: Willy Tarreau <w@1wt.eu>