BUG/MEDIUM: hlua: Make sure we drain the output buffer when done.
In hlua_applet_tcp_fct(), drain the output buffer when the applet is done
running, every time we're called.
Overwise, there's a race condition, and the output buffer could be filled
after the applet ran, and as it is never cleared, the stream interface
will never be destroyed.

This should be backported to 1.8 and 1.7.

(cherry picked from commit 594c8c5015abb0d0c0a8d41fea573a2bcd63a351)
[wt: s / co_data(si_oc(si)) / si_ob(si)->o]
Signed-off-by: Willy Tarreau <w@1wt.eu>