BUG/MEDIUM: h2: fix risk of memory leak on malformated wrapped frames
While parsing a headers frame, if the frame is wrapped in the buffer
and needs to be unwrapped, it will be duplicated before being processed.
But if it contains certain combinations of invalid flags, the parser
returns without releasing the temporary buffer leading to a memory
leak.

This fix needs to be backported to 1.8.

(cherry picked from commit a0d11b6fd5b6cec3af9a70e38895a665a666ae80)
Signed-off-by: Willy Tarreau <w@1wt.eu>