BUG/MINOR: cli: make sure the "getsock" command is only called on connections
Theorically nothing would prevent a front applet form connecting to a stats
socket, and if a "getsock" command was issued, it would cause a crash. Right
now nothing in the code does this so in its current form there is no impact.

It may or may not be backported to 1.8.

(cherry picked from commit c2b7f80a91ed01f78aa22b3e2afeda0672501e0c)
Signed-off-by: Willy Tarreau <w@1wt.eu>