BUG/MEDIUM: threads: make sure threads_want_sync is marked volatile
The threads_want_sync variable is not volatile, which allows the compiler
to cache old copies of it for long parts of code and possibly optimize
some tests away. This could result in deadlocks when using heavy queue
activity or health check state changes.

There is no upstream commit for this fix because the sync point was
completely removed from 1.9. This fix is exclusively for 1.8.

Signed-off-by: Willy Tarreau <w@1wt.eu>