DOC: Fix typos in README and CONTRIBUTING
Few typos detected by misspell in the README and CONTRIBUTING.
Even if one of them is on a listing of commits. I'm assuming that
if we want to enforce less typos in the commits, having one in the
contributing guide is not the best example.

(cherry picked from commit e07bc14e350a32a9ccf9e3d415e7d60c4d5bef09)
Signed-off-by: Willy Tarreau <w@1wt.eu>