BUG/MINOR: config: Copy default error messages when parsing of a backend starts
To be used, error messages declared in a default section must be copied when the
parsing of a proxy section starts. But this was only done for frontends.

This patch may be backported to older versions.

(cherry picked from commit 6b44975fbd2f7d81074e20bc07fc0e01466cc9c9)
Signed-off-by: William Lallemand <wlallemand@haproxy.org>