BUG/MEDIUM: hpack: fix encoding of "accept-ranges" field
James Brown reported that when an "accept-ranges" header field is sent
through haproxy and converted from HTTP/1.1 to H2, it's mis-encoded as
"accept-language". It happens that it's one of the few very common header
fields encoded using its index value and that this index value was misread
in the spec as 17 instead of 18, resulting in the wrong name being sent.
Thanks to Lukas for spotting the issue in the HPACK encoder itself.

This fix must be backported to 1.8.

(cherry picked from commit 4bf194cbdbcda8ec4ce83d7f12d2fe9b08483c94)
[wla: buffer API edit]
Signed-off-by: William Lallemand <wlallemand@haproxy.org>