BUG/MINOR: cfgparse: Fix the call to post parser of the last sections parsed
Wrong variable was used to know if we need to call the callback
post_section_parser() or not. We must use 'cs' and not 'pcs'.

This patch must be backported in 1.8 with the commit 7805e2b ("BUG/MINOR:
cfgparse: Fix transition between 2 sections with the same name").

(cherry picked from commit bf7a9597e2e27f999562043eca6b95763d166130)
Signed-off-by: Willy Tarreau <w@1wt.eu>