BUG/MINOR: lb-map: fix unprotected update to server's score
The loop trying to figure the best server is theorically capable of
finishing the loop with best == NULL, causing the HA_ATOMIC_SUB()
to fail there. However for this to happen the list should be empty,
which is avoided at the beginning of the function. As it is, the
function still remains at risk so better address this now.

This patch should be backported to 1.8.

(cherry picked from commit c8b476d9521e4355825ae89c24c12ffbbb1791c7)
Signed-off-by: Willy Tarreau <w@1wt.eu>