BUG: dns: Prevent stack-exhaustion via recursion loop in dns_read_name
When a compressed pointer is encountered, dns_read_name() will call
itself with the pointed-to offset in the packet.
With a specially crafted packet, it was possible to trigger an
infinite-loop recursion by making the pointer points to itself.
While it would be possible to handle that particular case differently
by making sure that the target is different from the current offset,
it would still be possible to craft a packet with a very long chain
of valid pointers, always pointing backwards. To prevent a stack
exhaustion in that case, this patch restricts the number of recursive
calls to 100, which should be more than enough.

To be backported to 1.8, probably also 1.7.

(cherry picked from commit 58df5aea0a0c926b2238f65908f5e9f83d1cca25)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 2b514b24f71af8ff8c6593636850b9a312a05278)
[wt: adjusted context]
Signed-off-by: Willy Tarreau <w@1wt.eu>