BUG: dns: Prevent out-of-bounds read in dns_read_name()
Some callers of dns_read_name() do not make sure that we can read
the first byte, holding the length of the next label, without going
past our buffer, so we need to make sure of that.
In addition, if the label is a compressed one we need to make sure
that we can read the following byte to compute the target offset.

To be backported to 1.8, probably also 1.7.

(cherry picked from commit 2d19fbcab21356f5aee1f2a52591a38665360bb6)
Signed-off-by: Willy Tarreau <w@1wt.eu>