BUG/MINOR: hpack: fix off-by-one in header name encoding length calculation
In hpack_encode_header() there is a length check to verify that a literal
header name fits in the buffer, but there it an off-by-one in this length
check, which forgets the byte required to mark the encoding type (literal
without indexing). It should be harmless though as it cannot be triggered
since response headers passing through haproxy are limited by the reserve,
which is not the case of the output buffer.

This fix should be backported to 1.8.

(cherry picked from commit 7571015939eb0ab560df0fefc434e273ec9d93ff)
[wt: adjust context]
Signed-off-by: Willy Tarreau <w@1wt.eu>