REGTEST: script: Add the option --keep-logs to keep all log directories
By default a log directory is kept only if the test fails. With this option it
is possible to always keep it. If this option is used, the result of all tests
are displayed (and not only the failing ones).

(cherry picked from commit 9c6df5ecb4acbf39e7413f97144ae767a7c2b414)
Signed-off-by: Willy Tarreau <w@1wt.eu>