BUG/MAJOR: connections: Close the connection before freeing it.
In si_release_endpoint(), if the end point is a connection, because we don't
know which mux to use it, make sure we close the connection before freeing it,
or else, we'd have a fd left for polling, which would point to a now free'd
connection.

This should be backported to 1.9.

(cherry picked from commit 8ab8a6eee5297c440e0a0dd1edec90a2def515f5)
Signed-off-by: Willy Tarreau <w@1wt.eu>