MINOR: buffers: add a new b_move() function
This function will be used to move parts of a buffer to another place
in the same buffer, even if the parts overlap. In order to keep things
under reasonable control, it only uses a length and absolute offsets
for the source and destination, and doesn't consider head nor data.

(cherry picked from commit f48919aafb3b1bb9120dce4d16c269dde530176f)
Signed-off-by: Willy Tarreau <w@1wt.eu>